Move DeleteDeviceButton to general folder
Since it's used in both devicelist and devicedetails now
