+++ frontend/src/devicelist/ActionMenu.jsx
@@ -5,7 +5,7 @@ import PropTypes from "prop-types";
 
 import { MoreVert } from "@material-ui/icons/";
 import DeviceInfo from "../model/DeviceInfo";
-import DeleteDeviceButton from "../deviceDetailsPage/DeleteDeviceButton";
+import DeleteDeviceButton from "../general/DeleteDeviceButton";
 
 export default class ActionMenu extends React.Component {
   constructor(props) {
