+++ frontend/src/general/DeleteDeviceButton.jsx
