+++ frontend/src/general/__tests__/DeleteDeviceButton.test.jsx
